backend/requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Environment variables
python-decouple==3.8

# Whisper and AI dependencies
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.3

# File handling
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
